// app/vokabeln/types.ts
export type Voc = {
  id: number;
  ord: number;
  art: string;
  artikel: string;
  wort: string;
  geschlecht: string;
  aussprache: string;
  uebersetzung: string;
  beispielsatz: string;
  beispielsatzuebersetzung: string;
  eigenerKommentar: string;
  sprache: string;
  book: string;
  chapter: string;
  schwierigkeitsgrad: number;
  kapitel: number;
  kapiteltext: string;
};